app_sequence/requirements.txt
streamlit==1.32.0
numpy==1.24.4
pandas==2.0.3
plotly==5.19.0
requests==2.31.0
Pillow==10.2.0
python-dotenv==1.0.1
streamlit-drawable-canvas==0.9.3
matplotlib==3.7.3 